refactor: HINT 객체 Messages로 이동
